requirements.txt
torch
opencv-python
numpy
matplotlib
Pillow
torchvision